pytest_attrib: Use the `parent` provided by pytest
Rather than meddling with the metadata ourselves, we now use the
`parent` attribute provided by pytest. This is much more
reliable as pytest searches for the testcases in a top
to bottom way (I.e. first find modules, then classes, then functions)
and so it becomes easy for it to track the heirarchy.