Move pytest_attrib into own module folder
Earlier whe using the version for setup.py the pytest_attrib
file needed to be imported. This is not right, because it imports
pytest, which may not be installed at this point. Now, we move it
into it's own folder and read the VERSION file for this.

Fixes https://github.com/AbdealiJK/pytest-attrib/issues/5